=별점시스템

+++ lib/screens/note_screen.dart
@@ -1,38 +1,104 @@
+import 'dart:js';
+
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:intl/intl.dart';
+import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 
 
 class NotePage extends StatelessWidget {
   final TextEditingController _postController = TextEditingController();
   final User? user = FirebaseAuth.instance.currentUser; // user 변수 선언 및 초기화
+  int selectedRating = 0;
 
-  Future<void> _addPost() async {
+  Future<void> _addPost(int selectedRating) async {
     if (user != null) {
       if (_postController.text.isNotEmpty) {
         await FirebaseFirestore.instance.collection('posts').add({
           'content': _postController.text.trim(),
           'timestamp': DateTime.now().toLocal(),
           'userId': user!.email,
+          'rating': selectedRating, // 추가된 부분
         });
         _postController.clear();
       }
     }
   }
 
-  Future<void> _deletePost(String documentId) async {
-    await FirebaseFirestore.instance.collection('posts')
-        .doc(documentId)
-        .delete();
+
+  Future<void> _deletePost(BuildContext context, String documentId, String postUserId) async {
+    if (user != null && user!.email == postUserId) {
+      await FirebaseFirestore.instance.collection('posts')
+          .doc(documentId)
+          .delete();
+    } else {
+      // 권한이 없는 경우 처리
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('글을 삭제할 권한이 없습니다.'),
+        ),
+      );
+    }
   }
+
+
 
 
   @override
   Widget build(BuildContext context) {
+    void _showReviewDialog() {
+      showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          int selectedRating = 0; // 사용자가 선택한 별점을 저장하는 변수
+
+          return AlertDialog(
+            title: Text('만족도 평가'),
+            content: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Text('별점을 선택하세요:'),
+                // 별점 선택 위젯
+                RatingBar.builder(
+                  initialRating: 0,
+                  minRating: 1,
+                  direction: Axis.horizontal,
+                  allowHalfRating: false,
+                  itemCount: 5,
+                  itemSize: 30.0,
+                  itemBuilder: (context, _) => Icon(
+                    Icons.star,
+                    color: Colors.amber,
+                  ),
+                  onRatingUpdate: (rating) {
+                    selectedRating = rating.toInt();
+                  },
+                ),
+              ],
+            ),
+            actions: [
+              TextButton(
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+                child: Text('취소'),
+              ),
+              TextButton(
+                onPressed: () {
+                  _addPost(selectedRating);
+                  Navigator.pop(context);
+                },
+                child: Text('작성'),
+              ),
+            ],
+          );
+        },
+      );
+    }
     return Scaffold(
       appBar: AppBar(
-        title: Text('게시판'),
+        title: Text('리뷰 게시판'),
       ),
       body: Column(
         children: [
@@ -57,14 +123,26 @@ class NotePage extends StatelessWidget {
                     Map<String, dynamic> data = document.data()! as Map<String, dynamic>;
                     return ListTile(
                       title: Text(data['content']),
-                      subtitle: Text(DateFormat('yyyy-MM-dd HH:mm:ss').format(data['timestamp'].toDate())),
+                      subtitle: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Text(DateFormat('yyyy-MM-dd HH:mm:ss').format(data['timestamp'].toDate())),
+                          // 별점 표시
+                          Row(
+                            children: [
+                              Icon(Icons.star, color: Colors.amber),
+                              Text(' ${data['rating'] ?? 'N/A'}'), // 별점이 없는 경우 'N/A'를 표시
+                            ],
+                          ),
+                        ],
+                      ),
                       trailing: IconButton(
                         icon: Icon(Icons.delete),
-                        onPressed: () => _deletePost(document.id),
+                        onPressed: () => _deletePost(context, document.id, data['userId']),
                       ),
-                      // 작성자 표시
-                      leading: Text(user?.email ?? 'Unknown'),
+                      leading: Text(data['userId'] ?? 'Unknown'),
                     );
+
                   },
                 );
               },
@@ -81,10 +159,12 @@ class NotePage extends StatelessWidget {
           ),
           ElevatedButton(
             child: Text('작성'),
-            onPressed: _addPost,
+            onPressed: () => _showReviewDialog(),
           ),
+
         ],
       ),
     );
   }
+
 }
